Make load tests executable as stand-alone application
Move code from test to main. Make code runnable with sbt run command.
Remove unnecessary config files.
