+++ load-test/.gitignore
@@ -1,6 +1,8 @@
 target/
 .classpath
 .cache-tests
+.cache-main
 .settings/
 .project
 *.class
+bin/
