+++ load-test/src/main/scala/com/mainflux/loadtest/simulations/PublishSimulation.scala
@@ -1,0 +1,106 @@
+package com.mainflux.loadtest.simulations
+
+import scala.concurrent.duration._
+import scalaj.http.Http
+import io.gatling.core.Predef._
+import io.gatling.http.Predef._
+import io.gatling.jdbc.Predef._
+import io.circe._
+import io.circe.generic.auto._
+import io.circe.parser._
+import io.circe.syntax._
+import PublishSimulation._
+import io.gatling.http.protocol.HttpProtocolBuilder.toHttpProtocol
+import io.gatling.http.request.builder.HttpRequestBuilder.toActionBuilder
+import com.mainflux.loadtest.Engine._
+
+class PublishSimulation extends Simulation {
+
+  // Register user
+  Http(s"${ManagerUrl}/users")
+    .postData(User)
+    .header(HttpHeaderNames.ContentType, ContentType)
+    .asString
+
+  // Login user
+  val tokenRes = Http(s"${ManagerUrl}/tokens")
+    .postData(User)
+    .header(HttpHeaderNames.ContentType, ContentType)
+    .asString
+    .body
+
+  val tokenCursor = parse(tokenRes).getOrElse(Json.Null).hcursor
+  val token = tokenCursor.downField("token").as[String].getOrElse("")
+
+  // Register client
+  val clientLocation = Http(s"${ManagerUrl}/clients")
+    .postData(Client)
+    .header(HttpHeaderNames.Authorization, token)
+    .header(HttpHeaderNames.ContentType, ContentType)
+    .asString
+    .headers.get("Location").get(0)
+
+  val clientId = clientLocation.split("/")(2)
+
+  // Get client key
+  val clientRes = Http(s"${ManagerUrl}/clients/${clientId}")
+    .header(HttpHeaderNames.Authorization, token)
+    .header(HttpHeaderNames.ContentType, ContentType)
+    .asString
+    .body
+
+  val clientCursor = parse(clientRes).getOrElse(Json.Null).hcursor
+  val clientKey = clientCursor.downField("key").as[String].getOrElse("")
+
+  // Register channel
+  val chanLocation = Http(s"${ManagerUrl}/channels")
+    .postData(Channel)
+    .header(HttpHeaderNames.Authorization, token)
+    .header(HttpHeaderNames.ContentType, ContentType)
+    .asString
+    .headers.get("Location").get(0)
+
+  val chanId = chanLocation.split("/")(2)
+
+  // Connect client to channel
+  Http(s"${ManagerUrl}/channels/${chanId}/clients/${clientId}")
+    .method("PUT")
+    .header(HttpHeaderNames.Authorization, token)
+    .asString
+
+  // Prepare testing scenario
+  val httpProtocol = http
+    .baseURL(HttpAdapterUrl)
+    .inferHtmlResources()
+    .acceptHeader("*/*")
+    .contentTypeHeader("application/json; charset=utf-8")
+    .userAgentHeader("curl/7.54.0")
+
+  val scn = scenario("PublishMessage")
+    .exec(http("request_0")
+      .post(s"/channels/${chanId}/messages")
+      .header(HttpHeaderNames.ContentType, "application/senml+json")
+      .header(HttpHeaderNames.Authorization, clientKey)
+      .body(StringBody(Message))
+      .check(status.is(202)))
+
+  setUp(
+    scn.inject(
+      constantUsersPerSec(100) during (15 second),
+      nothingFor(15 second),
+      constantUsersPerSec(250) during (15 second),
+      nothingFor(15 second),
+      constantUsersPerSec(500) during (15 second),
+      nothingFor(15 second),
+      constantUsersPerSec(750) during (15 second),
+      nothingFor(15 second),
+      constantUsersPerSec(1000) during (15 second))).protocols(httpProtocol)
+}
+
+object PublishSimulation {
+  val ContentType = "application/json; charset=utf-8"
+  val User = """{"email":"john.doe@email.com", "password":"123"}"""
+  val Client = """{"type":"device", "name":"weio"}"""
+  val Channel = """{"name":"mychan"}"""
+  val Message = """[{"bn":"some-base-name:","bt":1.276020076001e+09, "bu":"A","bver":5, "n":"voltage","u":"V","v":120.1}, {"n":"current","t":-5,"v":1.2}, {"n":"current","t":-4,"v":1.3}]"""
+}
