--- mainflux-loadtests/bin/bodies/CreateAndRetrieveClientSimulation_0000_request.txt
@@ -1,1 +1,0 @@
-{"type":"device", "name":"weio"}
