--- mainflux-loadtests/bin/bodies/PublishSimulation_0000_request.txt
@@ -1,1 +1,0 @@
-[{"bn":"some-base-name:","bt":1.276020076001e+09, "bu":"A","bver":5, "n":"voltage","u":"V","v":120.1}, {"n":"current","t":-5,"v":1.2}, {"n":"current","t":-4,"v":1.3}]
