--- mainflux-loadtests/project/plugins.sbt
@@ -1,1 +1,0 @@
-addSbtPlugin("io.gatling" % "gatling-sbt" % "2.2.2")
