--- mainflux-loadtests/src/test/scala/Recorder.scala
@@ -1,12 +1,0 @@
-import io.gatling.recorder.GatlingRecorder
-import io.gatling.recorder.config.RecorderPropertiesBuilder
-
-object Recorder extends App {
-
-	val props = new RecorderPropertiesBuilder
-	props.simulationOutputFolder(IDEPathHelper.recorderOutputDirectory.toString)
-	props.simulationPackage("com.mainflux.loadtests")
-	props.bodiesFolder(IDEPathHelper.bodiesDirectory.toString)
-
-	GatlingRecorder.fromMap(props.build, Some(IDEPathHelper.recorderConfigFile))
-}
